@@ -21,7 +21,7 @@ end
 # Note: This is so that slimmer is skipped, preventing network requests for
 # content from static (i.e. core_layout.html.erb).
 class ActionController::Base
-  before_filter proc {
+  before_action proc {
     response.headers[Slimmer::Headers::SKIP_HEADER] = "true" unless ENV["USE_SLIMMER"]
   }
 end
